requirements.txt
absl-py==2.1.0
accelerate==0.34.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs==24.2.0
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
comm @ file:///croot/comm_1709322850197/work
contourpy==1.3.0
cycler==0.12.1
datasets==2.21.0
debugpy @ file:///work/perseverance-python-buildout/croot/debugpy_1698884710808/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@ebe8b45437f86395352ab13402ba45b75b4d1ddb
dill==0.3.8
docker-pycreds==0.4.0
donut-python==1.0.9
evaluate==0.4.2
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.1
huggingface-hub==0.24.6
humanfriendly==10.0
hydra-core==1.3.2
idna==3.8
iopath==0.1.9
ipykernel @ file:///croot/ipykernel_1705933831282/work
ipython @ file:///croot/ipython_1718287989724/work
ipywidgets==8.1.5
jedi @ file:///croot/jedi_1721058342488/work
Jinja2==3.1.3
jiwer==3.0.5
joblib==1.4.2
jupyter_client @ file:///work/perseverance-python-buildout/croot/jupyter_client_1701731788627/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lanms-nova==1.0.3
lightning-utilities==0.11.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline @ file:///work/perseverance-python-buildout/croot/matplotlib-inline_1698864771271/work
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.2
opencv-python==4.10.0.84
optimum==1.22.0
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.2.2
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathspec==0.12.1
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow==10.2.0
pillow_heif==0.21.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
portalocker==2.10.1
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==5.28.1
psutil @ file:///work/perseverance-python-buildout/croot/psutil_1698863411559/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==17.0.0
pycocotools==2.0.8
Pygments @ file:///work/perseverance-python-buildout/croot/pygments_1698846270603/work
pyparsing==3.1.4
pytesseract==0.3.13
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML==6.0.2
pyzbar==0.1.9
pyzmq @ file:///croot/pyzmq_1705605076900/work
RapidFuzz==3.10.1
regex==2024.7.24
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
sconf==0.2.5
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.16.0
seqeval==1.2.2
setproctitle==1.3.3
setuptools==72.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.12
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
timm==1.0.10
tokenizers==0.20.1
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchinfo==1.8.0
torchmetrics==1.4.3
torchvision==0.19.1+cu118
tornado @ file:///croot/tornado_1718740109488/work
tqdm==4.66.5
traitlets @ file:///croot/traitlets_1718227057033/work
transformers==4.45.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.18.3
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==3.0.4
wheel==0.43.0
widgetsnbextension==4.0.13
xxhash==3.5.0
yacs==0.1.8
yarl==1.9.8
zss==1.2.0
